Add files via upload
In this strategy we are using the 5 and the 60 MA with the ratio. It is connecting to interactive brokers API and placing trades automatically. It is choosing from a Portfolio of assets to find the Co-integrated pair With the lowest P-score, Our goal, as always, is to help the little guy succeed, not the hedge funds.